pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "nonebot-plugin-hullqin-game"
version = "1.0.0"
description = "Hullqin Game 桌游发车"
authors = [
    {name = "Polaris_Light", email = "995905922@qq.com" }
]
dependencies = [
    "nonebot2>=2.2.1",
    "nonebot-adapter-onebot>=2.0.0-beta.1",
    "nonebot-plugin-localstore>=0.7.4",
    "pydantic>=1.10",
    "playwright>=1.37.0",
]
readme = "README.md"
license = "AGPL-3.0"
license-files = ["LICENSE*"]
requires-python = ">=3.10"
keywords = ["hullqin", "game", "桌游"]
classifiers = []

[project.urls]
Homepage = "https://github.com/GLDYM/nonebot-plugin-hullqin-game"

[tool.poetry.dependencies]
nonebot2 = ">=2.2.1"
nonebot-adapter-onebot = ">=2.0.0-beta.1"
nonebot-plugin-localstore = ">=0.7.4"
pydantic = ">=1.10"
playwright = ">=1.37.0"